Added different HUDs for weapon types
Set up HUD to change depending on weapon type set in the hitscan weapon script... this is set up for assault rifles, battle rifles, shotguns, sniper rifles and pistols. It should be set up to automatically work with newly added weapons without any setup. Some of the animations are a little janky and will need to be fixed in the future.